move binaries tests to another file
